feat-walk/refactor: 다이어리 작성 버튼 ui와 로직 분리
